+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 ### Version 7.6
+- Added option to delete an IFRA entry
+- IFRA Library extended to display implementation deadlines for new and current creations
+
+
 ### Version 7.5
 - Making sure dilutant reset to its default if previously used
 - Making scale formula modal static